Index paths using Boost R-tree
This replaces the previous method of storing all paths
in a vector and performing a linear search. Now, the
segments of each path are laid out over an R-tree which
facilitates faster retrieval when the cache is large.

In addition to the index, a primary reference to the paths
exists to allow optimization to take place. That collection
is now a list instead of a vector to reduce memory consumption
during the pruning phase.

+++ src/path/shareablepath.cpp
@@ -39,7 +39,7 @@ std::shared_ptr<Path> ShareablePath::PathBuilder::build() {
     if (!foundDestination || !foundOrigin) return nullptr;
 
     std::shared_ptr<Path> path = std::make_shared<Path2D>(subpathNodes[0], subpathNodes[1]);
-    for (int i = 2; i < subpathNodes.size(); i++) {
+    for (size_t i = 2; i < subpathNodes.size(); i++) {
         path->add(subpathNodes[i]);
     }
     return path;
