+++ src/provider/provideradapter.cpp
@@ -27,7 +27,7 @@ std::shared_ptr<Path> ProviderAdapter::convert(const std::vector<float> &rawPath
     
     Node a(rawPath[0], rawPath[1]), b(rawPath[2], rawPath[3]);
     std::shared_ptr<Path> path = std::make_shared<Path2D>(a, b);
-    for (int i = 4; i < coordCount - 1; i++) {
+    for (size_t i = 4; i < coordCount - 1; i++) {
         path->add(Node(rawPath[i], rawPath[i + 1]));
     }
     return path;
